+++ src/__mocks__/handlers.ts
@@ -2,11 +2,14 @@ import { rest } from 'msw';
 import querystring from 'querystring';
 const authUrl = `http://localhost/auth-oauth2/oauth/token`;
 
-const goodAuthResponseBody = {
-  access_token: 'proper_access_token',
+import jwt from 'jsonwebtoken';
+import dayjs from 'dayjs';
+
+export const goodAuthResponseBody = {
+  access_token: jwt.sign({ exp: dayjs().add(1, 'hour').unix() }, 'secret'),
   expires_in: '3600',
   UUID: 'user_uuid',
-  refresh_token: 'proper_refresh_token',
+  refresh_token: jwt.sign({ exp: dayjs().add(24, 'hour').unix() }, 'secret'),
 };
 
 export const handlers = [
@@ -32,7 +35,7 @@ export const handlers = [
     }
     if (grantType === 'refresh_token') {
       const refreshToken = params['refresh_token'];
-      if (refreshToken === 'proper_refresh_token') {
+      if (refreshToken === goodAuthResponseBody.refresh_token) {
         return resp(
           ctx.delay(2000),
           ctx.status(200),